Major refactoring.
Moved Padre::Frame to Padre::Wx::MainWindow.

Moved the frame from a global into the app object.

Refactored the wx test so it works just with the official objects and doesn't have to cheat.

Moved a number of modules inside of Class::Autouse loaders, so reduce the amount of code needed to start Padre, and thus make startup faster.

git-svn-id: file:///home/jurosz/dalsi/padre-svn-git/svn-mirror/trunk@83 18eddd70-d622-4293-bd05-fe65668e8bde