Landing the second half of the MainWindow destructive refactoring. I'd have liked to do many of these separately, but many of them kept overlapping with each other, so I've dropped it in one go. Changes include, but are not limited to...
Heavily changed (incomplete) Outline tool.
Heavily changed Syntax Checker.
The end of the ->{gui} hash.
Stable accessors to all tools are available now.
Changed all variables refering to a Padre::Wx::Main object to be $main.
Changed most variables refering to an Padre::Wx::Editor to be $editor.
Changed all show/hide logic to the new way used by the Function List.

(pauses to breath in...)

Renamed a ton of event handlers to match the name of the events they handle.
Padre::Current->main and Padre::Current->notebook are now public.
Changed most variables refering to a Padre::Wx::Notebook to $notebook.
Changed the naming scheme of the Timer identifiers in Padre::Wx to something that's actually decypherable.
Changed a ton of names in the AUI manager like bottompane or subs_panel to simpler ones that match their classes like bottom and functions.
Force the syntax checker to always be off when starting Padre (and never configurable), running arbitrary code automatically is an exploit waiting to happen.
Ensure Lock User Interface settings are applied when starting up, after the AuiManager loads it's perspective.

git-svn-id: file:///home/jurosz/dalsi/padre-svn-git/svn-mirror/trunk/Padre@2705 18eddd70-d622-4293-bd05-fe65668e8bde