Adding three new configutation settings options.
project: 'Can this setting by overwritten by Project policy settings'

options: An orders set of key value pairs describing the allowed options for this setting. The key is the setting enum, and value is a _T() localisable version for use in drop boxes and labels and the like.

apply: A CODE reference to a function that can be used to apply a new setting value to a currently-running live Padre install. This is added to support changing Padre configuration settings on the fly in the internals (without need to use a specific menu entry or config file).

git-svn-id: file:///home/jurosz/dalsi/padre-svn-git/svn-mirror/trunk@5767 18eddd70-d622-4293-bd05-fe65668e8bde