static.css the style="" attributes and fix download.html in large fonts.
I've spent a lot of time debugging the style="" attributes in the HTML
which over-rided all my changes to the stylesheet. I got rid of them
and replaced them with saner ones in the static.css.

I also fixed the fact that when resizing fonts the right column sometimes
overflowed to be placed below the left one.


git-svn-id: file:///home/jurosz/dalsi/padre-svn-git/svn-mirror/trunk@7595 18eddd70-d622-4293-bd05-fe65668e8bde