Implemented a state revisioning system to the View role for GUI elements.
This allows panel tools to indicate they have changed state and that any task result for the previous state should be discarded.

Added a View subclass for tasks which simplifies the process of writing tasks that will update a View based on their response, and that integrates with the state revisioning logic properly.

git-svn-id: file:///home/jurosz/dalsi/padre-svn-git/svn-mirror/branches/Padre-TaskRewrite@11513 18eddd70-d622-4293-bd05-fe65668e8bde