More refactoring to the owner/functionlist task stuff.
Now we can have a generic Padre::Task2::FunctionList task, with a document type specific one living at Padre::Document::Perl::FunctionList (and the code inlined).

Because the task is more focused and doesn't need to load the entire document type any more, the thread should be faster and smaller.

git-svn-id: file:///home/jurosz/dalsi/padre-svn-git/svn-mirror/branches/Padre-TaskRewrite@11528 18eddd70-d622-4293-bd05-fe65668e8bde