This commit really needs to be checked that it's doing the right thing!
The test failures were  "No Warnings" fails.  This was due to the path and name values not being returned.

The change here is to simply set an empty string rather than undef.
 
Testing the change doesn't seem to have broken anything that i can see, but I'm not convinced this is what you want when you have a directory listing.

This change though does mean all tests pass again.  By no means is this a recommendation this change isn't reviewed closely.

:)

git-svn-id: file:///home/jurosz/dalsi/padre-svn-git/svn-mirror/trunk/Padre@11725 18eddd70-d622-4293-bd05-fe65668e8bde