!!!MAJOR BUG FIX!!!
When the TaskManager attempts to spawn a new worker, while also inside of an active "DB" lock (via Padre::Locker) the database needs to be disconnected at the moment of the thread spawn.

The code that reconnects to the database uses the non-existant method Padre::DB->begin_work instead of Padre::DB->begin, resulting in corrupt TaskManager state and a failed unlock.

git-svn-id: file:///home/jurosz/dalsi/padre-svn-git/svn-mirror/trunk@13823 18eddd70-d622-4293-bd05-fe65668e8bde