on_right_click => on_context_menu
A first step to go to a clean support of the context menu key (available on
Windows keyboards) and its alternatives (right click or Shift+F10) in the editor.
But on GTK at least, EVENT_CONTEXT_MENU does not seem to work in Scintilla. So
we still use EVENT_RIGHT_CLICK to trigger the menu.


git-svn-id: file:///home/jurosz/dalsi/padre-svn-git/svn-mirror/trunk@15822 18eddd70-d622-4293-bd05-fe65668e8bde